+++ markdown/postprocessor/postprocessor.go
@@ -17,6 +17,7 @@ func Process(text string, url string) string {
 	text = filterSite(text, url)
 	text = moveZeroWidthSpaceUpOneLine(text)
 	text = indent(text)
+	text = deIndentInfoSection(text)
 
 	return text
 }
@@ -33,4 +34,27 @@ func indent(commentSection string) string {
 	indentedCommentSection, _ := t.WrapWithPad(commentSection, screenWidth, indentBlock)
 
 	return indentedCommentSection
+}
+
+func deIndentInfoSection(commentSection string) string {
+	var sb strings.Builder
+
+	lines := strings.Split(commentSection, "\n")
+
+	for _, line := range lines {
+		isInfoSection := strings.Contains(line, "╭") || strings.Contains(line, "│") ||
+			strings.Contains(line, "╰")
+
+		if isInfoSection {
+			deIndentedLine := strings.TrimPrefix(line, " ")
+
+			sb.WriteString(deIndentedLine + "\n")
+
+			continue
+		}
+
+		sb.WriteString(line + "\n")
+	}
+
+	return sb.String()
 }