+++ meta/meta.go
@@ -20,9 +20,27 @@ const (
 	newParagraph = "\n\n"
 )
 
+func GetReaderModeMetaBlock(title string, url string, lineWidth int) string {
+	style := lipgloss.NewStyle().
+		BorderStyle(lipgloss.RoundedBorder()).
+		PaddingLeft(1).
+		PaddingRight(1).
+		Width(lineWidth)
+
+	formattedTitle, _ := text.Wrap(Bold(title).String(), lineWidth)
+	formattedTitle = unicode.ZeroWidthSpace + newLine + formattedTitle
+	formattedUrl := Blue(text.TruncateMax(url, lineWidth-2)).String()
+	info := newParagraph + Green("Reader Mode").String()
+
+	// formattedTitle := Bold(title).String()
+	// formattedUrl := getURL(url, "_", lineWidth)
+
+	return formattedTitle + newParagraph + style.Render(formattedUrl+info) + newParagraph
+}
+
 func GetCommentSectionMetaBlock(c *item.Item, config *core.Config) string {
 	columnWidth := config.CommentWidth/2 - 1
-	url := getURL(c.URL, c.Domain, config)
+	url := getURL(c.URL, c.Domain, config.CommentWidth)
 	rootComment := parseRootComment(c.Content, config)
 
 	style := lipgloss.NewStyle().
@@ -75,12 +93,12 @@ func highlightTitle(title string, highlightHeadlines bool) string {
 	return Bold(highlightedTitle).String()
 }
 
-func getURL(url string, domain string, config *core.Config) string {
+func getURL(url string, domain string, lineWidth int) string {
 	if domain == "" {
 		return ""
 	}
 
-	truncatedURL := text.TruncateMax(url, config.CommentWidth-2)
+	truncatedURL := text.TruncateMax(url, lineWidth-2)
 	formattedURL := Blue(truncatedURL).String() + newLine
 
 	return formattedURL + newLine